Refactors homepage trending wikis code (#1117)
